fix: change names files, folder and component, delete one excess import React, rename example => UI
