+++ src/components/ui/tag/tag.stories.tsx
@@ -1,0 +1,15 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+
+import { Tag } from './tag';
+
+export default {
+  title: 'UI/Tag',
+  component: Tag,
+} as ComponentMeta<typeof Tag>;
+
+const Template: ComponentStory<typeof Tag> = (args) => <Tag {...args} />;
+
+export const Example = Template.bind({});
+Example.args = {
+  label: 'внеконкурсная программа'
+};
