--- src/components/ui/tags/index.ts
@@ -1,1 +1,0 @@
-export * from './tags';
